Ints all to int32: errors between spark and vastdb-sdk handling in table

+++ colabfit/tools/utilities.py
@@ -118,7 +118,7 @@ def spark_to_arrow_type(spark_type):
     if isinstance(spark_type, IntegerType):
         return pa.int32()
     elif isinstance(spark_type, LongType):
-        return pa.int64()
+        return pa.int32()
     elif isinstance(spark_type, DoubleType):
         return pa.float64()
     elif isinstance(spark_type, StringType):
